remove functions from pose estimating app to lib
